+++ src/main/java/roomescape/controller/dto/request/ThemeCreateRequest.java
@@ -1,0 +1,10 @@
+package roomescape.controller.dto.request;
+
+import roomescape.service.dto.input.ThemeInput;
+
+public record ThemeCreateRequest(String name, String description, String thumbnail) {
+
+    public ThemeInput toInput() {
+        return new ThemeInput(name, description, thumbnail);
+    }
+}
